Api - Adição de itens e correções.

+++ 01-Applicacao/Simasoft.Challenge.Lucro.Aplicacao/Dto/DistribuicaoLucro/ResultadoDistribuicaoLucrosDto.cs
@@ -5,9 +5,9 @@ namespace Simasoft.Challenge.Lucro.Aplicacao.Dto.DistribuicaoLucro
     public class ResultadoDistribuicaoLucrosDto
     {
         public float TotalFuncionarios { get; set; }
-        public float TotalDistribuido { get; set; }
-        public float TotalDisponibilizado { get; set; }
-        public float SaldoTotalDisponibilizado { get; set; }
+        public decimal TotalDistribuido { get; set; }
+        public decimal TotalDisponibilizado { get; set; }
+        public decimal SaldoTotalDisponibilizado { get; set; }
         public List<ParticipacaoDto> Participacoes { get; set; }
     }
 }
